- exceptions now derive from a common DamarisException class, which in turn derives from std::runtime_error   (previously, all exceptions derived from std::string, which is not a good idea) - I also added a RecoverableException class for ADC_result and the like to derive from. This class should be   used whenever code is caught inside the backend, e.g. in the experiment loop, and passed on to the frontend   without aborting - backends were modified accordingly - started adding a new test suite for more thorough unit tests and to understand program flow better - the working directory is now reset after a core is destroyed, also cleaned up the getcwd code

+++ damaris/backends/core/job.cpp
@@ -431,7 +431,7 @@ result* experiment::do_it(hardware* hw) {
   if (experiment_states==NULL) {
     return new error_result(job_no,"did not find any states in job file");
   }
-  fprintf(stderr,"performing experiment job no %" SIZETPRINTFLETTER ", number of exps is %i\n",job_no, experiment_states->size());
+  fprintf(stderr,"performing experiment job no %" SIZETPRINTFLETTER ", number of exps is %i\n",job_no, static_cast<int>(experiment_states->size()));
   result* data=hw->experiment(*experiment_states);
   fprintf(stderr,"finished experiment job no %" SIZETPRINTFLETTER "\n\n",job_no);
   if (data==NULL)
